+++ src/services/ios/screen-capture/index.ts
@@ -3,20 +3,10 @@ import { DVTSecureSocketProxyService } from '../dvt/index.js';
 import { Screenshot } from '../dvt/instruments/screenshot.js';
 import {
   ScreenCaptureStreamer,
+  type CaptureScreenshotOptions,
+  type CaptureScreenshotResult,
   type ScreenCaptureStreamerOptions,
 } from './streamer.js';
-
-export interface CaptureScreenshotOptions {
-  /** DVT screenshot captures the primary display; this is returned as metadata only. */
-  displayUniqueId?: string | null;
-}
-
-export interface CaptureScreenshotResult {
-  image: Buffer;
-  displayUniqueID?: string | null;
-  imageFormat?: string;
-  [key: string]: unknown;
-}
 
 export class ScreenCaptureService extends BaseService {
   static readonly RSD_SERVICE_NAME =
@@ -37,7 +27,12 @@ export class ScreenCaptureService extends BaseService {
     options: CaptureScreenshotOptions = {},
   ): Promise<CaptureScreenshotResult> {
     const screenshot = await this.getScreenshotInstrument();
-    return await this.captureWithInstrument(screenshot, options);
+    const image = await screenshot.getScreenshot();
+    return {
+      image,
+      displayUniqueID: options.displayUniqueId ?? null,
+      imageFormat: 'png',
+    };
   }
 
   createStreamer(options: ScreenCaptureStreamerOptions): ScreenCaptureStreamer {
@@ -47,7 +42,7 @@ export class ScreenCaptureService extends BaseService {
 
     this.activeStreamer = new ScreenCaptureStreamer(
       options,
-      this.createFrameCapture(options),
+      async (captureOptions) => await this.captureScreenshot(captureOptions),
       (streamer) => {
         if (this.activeStreamer === streamer) {
           this.activeStreamer = null;
@@ -81,35 +76,13 @@ export class ScreenCaptureService extends BaseService {
     await dvtService.connect();
     this.dvtService = dvtService;
     this.screenshot = new Screenshot(dvtService);
-    await this.screenshot.initialize();
     return this.screenshot;
-  }
-
-  private createFrameCapture(
-    options: CaptureScreenshotOptions,
-  ): () => Promise<CaptureScreenshotResult> {
-    const captureOptions = {
-      displayUniqueId: options.displayUniqueId,
-    };
-    let screenshot: Screenshot | null = null;
-
-    return async () => {
-      screenshot ??= await this.getScreenshotInstrument();
-      return await this.captureWithInstrument(screenshot, captureOptions);
-    };
-  }
-
-  private async captureWithInstrument(
-    screenshot: Screenshot,
-    options: CaptureScreenshotOptions,
-  ): Promise<CaptureScreenshotResult> {
-    const image = await screenshot.takeScreenshot();
-    return {
-      image,
-      displayUniqueID: options.displayUniqueId ?? null,
-      imageFormat: 'png',
-    };
   }
 }
 
-export type { ScreenCaptureStreamer, ScreenCaptureStreamerOptions };
+export type {
+  CaptureScreenshotOptions,
+  CaptureScreenshotResult,
+  ScreenCaptureStreamer,
+  ScreenCaptureStreamerOptions,
+};
